added scheleton for novosparc integration
