+++ spacemake/spatial/__init__.py
@@ -1,0 +1,3 @@
+# include in top level for backward compatibility
+from .util import compute_neighbors, compute_islands, detect_tissue, \
+    create_mesh, create_meshed_adata
